+++ CHANGELOG.md
@@ -7,6 +7,18 @@ The format is based on
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) this project adheres
 to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Change
+- Change command names to match built-in commands (old names can be used for compatibility but will be removed in future):
+  |            Old Command Names            |                New Command Names                |
+  | --------------------------------------- | ----------------------------------------------- |
+  | `extension.cursorNextWordEndJa`         | `japaneseWordHandler.cursorWordEndRight`        |
+  | `extension.cursorNextWordEndSelectJa`   | `japaneseWordHandler.cursorWordEndRightSelect`  |
+  | `extension.cursorPrevWordStartJa`       | `japaneseWordHandler.cursorWordStartLeft`       |
+  | `extension.cursorPrevWordStartSelectJa` | `japaneseWordHandler.cursorWordStartLeftSelect` |
+  | `extension.deleteWordRight`             | `japaneseWordHandler.deleteWordRight`           |
+  | `extension.deleteWordLeft`              | `japaneseWordHandler.deleteWordLeft`            |
+
 ## [1.1.1] - 2018-10-17
 ### Fixed
 - Now scrolls the window so that the cursor is always visible except when there are multiple cursors